* 3547_cleanup: (61 commits)
  Update po/mc.pot.
  Remove unneeded `struct` keyword for typedef'd structs
  src/editor/edit.c: Cleanup some compiler warnings
  src/editor/editbuffer.c: Cleanup some compiler warning
  src/editor/editwidget.c: Cleanup some compiler warnings
  (etags_set_definition_hash): refactoring:
  Add -Wassign-enum compiler wirning.
  (widget_options_t): add W_DEFAULT value to cleanup -Wassign-enum warnings.
  Add three new compiler warnings:
  Cleanup -Wfloat-conversion warning.
  (edit_block_copy_cmd): cleanup -Wconditional-uninitialized warning.
  (edit_search_fix_search_start_if_selection): cleanup -Wlogical-not-parentheses warning.
  (mc_error_message): make output format translatable.
  Find File: clarify description of exit from panelize mode.
  WPanel should report MSG_NOT_HANDLED for unhandled commands.
  Sync with gnulib 781788eeb5d4d0b816698e8629300cd90a7379bd.
  lib/widget/input.c: (quote): change type to gboolean.
  (tree_chdir, sync_tree): take vfs_path_t object to avoid double conversion of path.
  (set_display_type): ret rid of string reduplication.
  (expand_format): ret rid of extra string duplication.
  ...

+++ src/filemanager/filegui.c
@@ -343,9 +343,9 @@ static void
 file_frmt_time (char *buffer, double eta_secs)
 {
     int eta_hours, eta_mins, eta_s;
-    eta_hours = eta_secs / (60 * 60);
-    eta_mins = (eta_secs - (eta_hours * 60 * 60)) / 60;
-    eta_s = eta_secs - (eta_hours * 60 * 60 + eta_mins * 60);
+    eta_hours = (int) (eta_secs / (60 * 60));
+    eta_mins = (int) ((eta_secs - (eta_hours * 60 * 60)) / 60);
+    eta_s = (int) (eta_secs - (eta_hours * 60 * 60 + eta_mins * 60));
     g_snprintf (buffer, BUF_TINY, _("%d:%02d.%02d"), eta_hours, eta_mins, eta_s);
 }
